+++ packages/target-electron/src/ipc.ts
@@ -12,7 +12,6 @@ import path, { basename, extname, join, posix, sep, dirname } from 'path'
 import { inspect } from 'util'
 import { platform } from 'os'
 import { existsSync } from 'fs'
-import mimeTypes from 'mime-types'
 import { versions } from 'process'
 import { fileURLToPath } from 'url'
 
@@ -248,9 +247,6 @@ export async function init(cwd: string, logHandler: LogHandler) {
     }
   )
 
-  ipcMain.handle('app.writeClipboardToTempFile', () =>
-    writeClipboardToTempFile()
-  )
   ipcMain.handle('app.writeTempFileFromBase64', (_ev, name, content) =>
     writeTempFileFromBase64(name, content)
   )
@@ -341,26 +337,6 @@ export async function init(cwd: string, logHandler: LogHandler) {
     // the shutdown function
     dcController.jsonrpcRemote.rpc.stopIoForAllAccounts()
   }
-}
-
-async function writeClipboardToTempFile(): Promise<string> {
-  await mkdir(getDraftTempDir(), { recursive: true })
-  const formats = clipboard.availableFormats().sort()
-  log.debug('Clipboard available formats:', formats)
-  if (formats.length <= 0) {
-    throw new Error('No files to write')
-  }
-  const pathToFile = join(
-    getDraftTempDir(),
-    `paste.${mimeTypes.extension(formats[0]) || 'bin'}`
-  )
-  const buf =
-    mimeTypes.extension(formats[0]) === 'png'
-      ? clipboard.readImage().toPNG()
-      : clipboard.readBuffer(formats[0])
-  log.debug(`Writing clipboard ${formats[0]} to file ${pathToFile}`)
-  await writeFile(pathToFile, buf, 'binary')
-  return pathToFile
 }
 
 export async function writeTempFileFromBase64(